backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "FastAPI backend"
authors = ["Kiro"]
readme = "README.md"
packages = [{ include = "app", from = "src" }]

[tool.poetry.dependencies]
python  = ">=3.10,<4.0"
fastapi = ">=0.120.4,<0.121.0"
uvicorn = ">=0.38.0,<0.39.0"
sqlmodel = "^0.0.27"
asyncpg = "^0.30.0"
pydantic-settings = "^2.11.0"
jwt = "^1.4.0"
pyjwt = ">=2.8"
requests = "^2.32.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.server:run"

[dependency-groups]
dev = [
    "pytest (>=9.0.0,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
